move QueryStore to @graphiql/toolkit
