+++ .changeset/curvy-news-pull.md
@@ -1,0 +1,6 @@
+---
+'graphiql': minor
+'@graphiql/toolkit': minor
+---
+
+Move QueryStore from `graphiql` package to `@graphiql/toolkit`
