+++ packages/graphiql-toolkit/src/storage/index.ts
@@ -1,1 +1,2 @@
 export * from './base';
+export * from './query';
