+++ packages/graphiql/src/utility/HistoryStore.ts
@@ -4,8 +4,7 @@
  *  This source code is licensed under the MIT license found in the
  *  LICENSE file in the root directory of this source tree.
  */
-import QueryStore, { QueryStoreItem } from './QueryStore';
-import { StorageAPI } from '@graphiql/toolkit';
+import { QueryStore, QueryStoreItem, StorageAPI } from '@graphiql/toolkit';
 import { parse } from 'graphql';
 import {
   HandleEditLabelFn,
